# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "LMS_Lexicon.Api",
    "version": "v1"
  },
  "paths": {
    "/api/Authors": {
      "get": {
        "tags": [
          "Authors"
        ],
        "parameters": [
          {
            "name": "includeLiterature",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Author"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Author"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Author"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Author"
                  }
                }
              },
              "text/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Author"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Authors"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/Author"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Author"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Author"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Author"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/Author"
              }
            },
            "text/xml": {
              "schema": {
                "$ref": "#/components/schemas/Author"
              }
            },
            "application/*+xml": {
              "schema": {
                "$ref": "#/components/schemas/Author"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Author"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Author"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Author"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Author"
                }
              },
              "text/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Author"
                }
              }
            }
          }
        }
      }
    },
    "/api/Authors/{id}": {
      "get": {
        "tags": [
          "Authors"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Author"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Author"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Author"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Author"
                }
              },
              "text/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Author"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Authors"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/Author"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Author"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Author"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Author"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/Author"
              }
            },
            "text/xml": {
              "schema": {
                "$ref": "#/components/schemas/Author"
              }
            },
            "application/*+xml": {
              "schema": {
                "$ref": "#/components/schemas/Author"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Authors"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Literatures": {
      "get": {
        "tags": [
          "Literatures"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Literature"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Literature"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Literature"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Literature"
                  }
                }
              },
              "text/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Literature"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Literatures"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/Literature"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Literature"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Literature"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Literature"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/Literature"
              }
            },
            "text/xml": {
              "schema": {
                "$ref": "#/components/schemas/Literature"
              }
            },
            "application/*+xml": {
              "schema": {
                "$ref": "#/components/schemas/Literature"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Literature"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Literature"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Literature"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Literature"
                }
              },
              "text/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Literature"
                }
              }
            }
          }
        }
      }
    },
    "/api/Literatures/{id}": {
      "get": {
        "tags": [
          "Literatures"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Literature"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Literature"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Literature"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Literature"
                }
              },
              "text/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Literature"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Literatures"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/Literature"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Literature"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Literature"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Literature"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/Literature"
              }
            },
            "text/xml": {
              "schema": {
                "$ref": "#/components/schemas/Literature"
              }
            },
            "application/*+xml": {
              "schema": {
                "$ref": "#/components/schemas/Literature"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Literatures"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Subject": {
        "type": "object",
        "properties": {
          "subjectId": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "literatures": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Literature"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Literature": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "publicationDate": {
            "type": "string",
            "format": "date-time"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "subjectId": {
            "type": "integer",
            "format": "int32"
          },
          "level": {
            "type": "integer",
            "format": "int32"
          },
          "authors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Author"
            },
            "nullable": true
          },
          "subject": {
            "$ref": "#/components/schemas/Subject"
          }
        },
        "additionalProperties": false
      },
      "Author": {
        "type": "object",
        "properties": {
          "authorId": {
            "type": "integer",
            "format": "int32"
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "birthDate": {
            "type": "string",
            "format": "date-time"
          },
          "literatures": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Literature"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}